Repository Concluído! :)

+++ BiomeLab/src/main/java/com/BiomeLab/Model/Ambiente.java
@@ -25,7 +25,7 @@ public class Ambiente {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_ambiente")
-    private Integer idAmbiente;
+    private Long idAmbiente;
 
     @Column(name = "nm_ambiente", nullable = false, length = 100)
     private String nomeAmbiente;
